lambda/pdf_converter/requirements.txt
boto3>=1.35
pdf2image>=1.17
Pillow>=11.0
